requirements.txt
asyncqt
awaits
logger
logzero
python-docx
configparser
SpeechRecognition
vosk
